fix(sfcgal): Fix solid to polyhedralsurface conversion
When creating a `PolyhedralSurface` from a `Solid` it is necessary to
loop through all its shells and copy them to the new
`Polygon`. `sfcgal_polyhedral_surface_polygon_n` is used to retrieve
the shells. However, it only works for `PolyhedralSurface` and it
returns an empty polygon in that case.

This issue is fixed by calling `sfcgal_solid_shell_n` which is
supposed to be used in that case for `Solid`.