fix: icontract decorators must have a lambda as the first parameter
icontract actually scans the module which calls its decorators, and
pass the corresponding code line as a string in order to evaluate the
decorator validity.

By overloading the decorator with our `cond_icontract`, we break this
behavior, as we define the decorator type as the first argument
("require", "ensure", ...).

Plus, icontract does not support multiple decorators of the same
nature (e.g. 2 "require" decorators does not seem possible). Then this
commit merge every set of decorator predicates in a single predicate
for each method.