chore(update_def): Switch to a python version and handle deprecated
This removes the `update_def.sh` bash file and replaces it with a
python version as suggested by Loïc Bartoletti.

This also adds a new functionality: the removal of the "deprecated"
functions. Indeed `SFCGAL` C API now deprecates some functions. There
functions are marked as `deprecated` with the appropriate
keyword. However, `cffi` is not able to handle this keyword and the
parsing fails. This issue is fixed by removing theses functions from
the generated c files.

Co-authored-by: Jean Felder <jean.felder@oslandia.com>